Create http client wrapper. Implement test for geoceode service.

+++ src/main/kotlin/pl/piotr/weatherstation/weather/domain/entity/HourlyWeather.kt
@@ -23,7 +23,7 @@ data class HourlyWeather(
     pm25: Double,
     pm10: Double,
     rainGauge: Double,
-    windSpeedMax: Float,
+    windSpeedMax: Double,
     windSpeedAvg: Double,
     hourOfDay: Int,
   ) : this(
